separating code and tests
